requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-bootstrap-icons==0.8.3
django-bootstrap-modal-forms==2.2.1
django-bootstrap5==23.1
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
urllib3==1.26.15
